requirements.txt
google-generativeai
openai
pymongo
python-socketio
requests
schedule
boto3
nltk
deepgram-sdk==3.*
